% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{healthyverse_conflicts}
\alias{healthyverse_conflicts}
\title{Conflicts between the healthyverse and other packages}
\usage{
healthyverse_conflicts()
}
\value{
A message in regards to library conflicts.
}
\description{
This function lists all the conflicts between packages in the healthyverse
and other packages that you have loaded.
}
\details{
There are four conflicts that are deliberately ignored: \code{intersect},
\code{union}, \code{setequal}, and \code{setdiff} from dplyr. These functions
make the base equivalents generic, so shouldn't negatively affect any
existing code.
}
\examples{
healthyverse_conflicts()
}
