% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{healthyverse_deps}
\alias{healthyverse_deps}
\title{List all healthyverse dependencies}
\usage{
healthyverse_deps(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
healthyverse packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOptions("repos")}.}
}
\value{
A message indicating the dependencies of the healthyverse
}
\description{
List all healthyverse dependencies
}
