% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{healthyverse_packages}
\alias{healthyverse_packages}
\title{List all packages in the healthyverse}
\usage{
healthyverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include healthyverse in the list?}
}
\value{
A list of the healthyverse packages
}
\description{
List all packages in the healthyverse
}
\examples{
healthyverse_packages()
}
