% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{healthyverse_update}
\alias{healthyverse_update}
\title{Update healthyverse packages}
\usage{
healthyverse_update(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
healthyverse packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOptions("repos")}.}
}
\value{
This function do not return a value but rather are used for side effects.
}
\description{
This will check to see if all healthyverse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
healthyverse_update()
}
}
